requirements.txt
asgiref==3.3.4
cx-Oracle==8.2.1
Django==3.2.4
django-db-file-storage==0.5.5
djangorestframework==3.12.4
Pillow==8.2.0
python-dotenv==0.17.1
pytz==2021.1
sqlparse==0.4.1
